Improve API structure and complete implementation
Refactor the chat API to improve structure and maintainability.

* **Add `src/controllers/chatController.ts`**:
  - Create a new file to handle chat-related logic.
  - Move event handlers from `src/routes/chat.ts` to this file.
  - Export functions for each event handler.
  - Import necessary models and utilities.
  - Ensure all functions are properly typed.

* **Add `src/services/chatService.ts`**:
  - Create a new file to handle chat-related services.
  - Move conversation and message-related logic from `src/routes/chat.ts` to this file.
  - Export functions for each service.
  - Import necessary models and utilities.
  - Ensure all functions are properly typed.

* **Modify `src/routes/chat.ts`**:
  - Remove event handlers and move them to `src/controllers/chatController.ts`.
  - Import and use functions from `src/controllers/chatController.ts`.
  - Ensure all routes are properly typed.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/OxyHQ/Allo?shareId=XXXX-XXXX-XXXX-XXXX).
