+++ .devcontainer/devcontainer.json
@@ -1,0 +1,5 @@
+{
+  "tasks": {
+    "build": "npm install && npm run build"
+  }
+}
